Refactor progress actions into modular structure
