+++ tests/gameLifecycle.test.js
@@ -11,7 +11,7 @@ const assetsModule = await import('../src/game/assets/index.js');
 const hustlesModule = await import('../src/game/hustles.js');
 const upgradesModule = await import('../src/game/upgrades.js');
 const requirementsModule = await import('../src/game/requirements.js');
-const actionsProgressModule = await import('../src/game/actions/progress.js');
+const actionsProgressModule = await import('../src/game/actions/progress/instances.js');
 const lifecycleModule = await import('../src/game/lifecycle.js');
 const hustleMarketSlice = await import('../src/core/state/slices/hustleMarket.js');
